requirements.txt
altair==5.3.0
appdirs==1.4.4
apturl==0.5.2
asttokens==2.4.1
attrs==23.2.0
Babel==2.8.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
blinker==1.8.2
Brlapi==0.8.3
Brotli==1.0.9
cachetools==5.3.3
certifi==2024.6.2
chardet==4.0.0
charset-normalizer==3.3.2
chrome-gnome-shell==0.0.0
click==8.1.7
colorama==0.4.4
comm==0.2.2
command-not-found==0.3
configobj==5.0.6
contourpy==1.2.1
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.18
debugpy==1.8.1
decorator==4.4.2
defer==1.0.6
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu0.2
dnspython==2.7.0
duplicity==0.8.21
emoji==2.12.1
et-xmlfile==1.1.0
exceptiongroup==1.2.1
executing==2.0.1
fasteners==0.14.1
filelock==3.6.0
fonttools==4.53.0
fs==2.4.12
future==0.18.2
gast==0.5.2
GDAL==3.4.1
gitdb==4.0.11
GitPython==3.1.43
html5lib==1.1
httplib2==0.20.2
idna==3.7
importlib-metadata==4.6.4
ipykernel==6.29.4
ipython==8.24.0
ipython_genutils==0.2.0
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
keyring==23.5.0
kiwisolver==1.4.5
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
monotonic==1.6
more-itertools==8.10.0
mpmath==0.0.0
nbformat==5.1.3
nest-asyncio==1.6.0
netifaces==0.11.0
numpy==1.26.4
oauthlib==3.2.0
olefile==0.46
openpyxl==3.1.5
OWSLib==0.25.0
packaging==24.1
pandas==2.2.2
paramiko==2.9.3
parso==0.8.4
pbr==5.8.0
pexpect==4.8.0
pillow==10.3.0
platformdirs==2.5.1
plotly==5.4.0
ply==3.11
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
psycopg2==2.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.1.0
pycairo==1.20.1
pycups==2.0.1
pydeck==0.9.1
Pygments==2.18.0
PyGObject==3.42.1
pyinotify==0.9.6
PyJWT==2.3.0
pymacaroons==0.13.0
pymongo==4.10.1
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==3.1.2
pyproj==3.3.0
PyQt5==5.15.6
PyQt5-sip==12.9.1
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-docx==1.1.2
python-dotenv==1.0.1
python-pam==1.8.4
python-xapp==2.2.1
python-xlib==0.29
pythran==0.10.0
pytz==2024.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==26.0.3
QScintilla==2.11.6
referencing==0.35.1
reportlab==3.6.8
requests==2.32.3
rich==13.7.1
rpds-py==0.18.1
scipy==1.8.0
scour==0.38.2
screen-resolution-extra==0.0.0
seaborn==0.13.2
SecretStorage==3.3.1
setproctitle==1.2.2
six==1.16.0
smmap==5.0.1
soupsieve==2.3.1
stack-data==0.6.3
streamlit==1.44.0
sympy==1.9
systemd-python==234
tenacity==8.3.0
tinycss==0.4
tinycss2==1.1.1
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
urllib3==2.2.1
usb-creator==0.3.7
variety==0.8.5
virtualenv==20.13.0+ds
wadllib==1.3.6
watchdog==4.0.1
wcwidth==0.2.13
webencodings==0.5.1
xdg==5
xkit==0.0.0
xlrd==1.2.0
zipp==1.0.0
